Refactor API file layout fixes #2693
