+++ app/experimenter/experiments/tests/test_forms.py
@@ -8,6 +8,7 @@ from faker import Factory as FakerFactory
 from django.contrib.auth.models import Permission
 from django.contrib.contenttypes.models import ContentType
 
+from experimenter.base.tests.factories import CountryFactory, LocaleFactory
 from experimenter.experiments.forms import (
     BugzillaURLField,
     ChangeLogMixin,
@@ -29,9 +30,7 @@ from experimenter.experiments.forms import (
 )
 from experimenter.experiments.models import Experiment
 from experimenter.experiments.tests.factories import (
-    CountryFactory,
     ExperimentFactory,
-    LocaleFactory,
     ProjectFactory,
     UserFactory,
 )
